requirements.txt
asgiref
autopep8
Django
django-cors-headers
djangorestframework
Pillow
pycodestyle
pytz
sqlparse
toml
markdown
django-tinymce


